+++ src/sagemaker/amazon/lda.py
@@ -13,7 +13,7 @@
 from sagemaker.amazon.amazon_estimator import AmazonAlgorithmEstimatorBase, registry
 from sagemaker.amazon.common import numpy_to_record_serializer, record_deserializer
 from sagemaker.amazon.hyperparameter import Hyperparameter as hp  # noqa
-from sagemaker.amazon.validation import gt, isint, isnumber
+from sagemaker.amazon.validation import gt
 from sagemaker.predictor import RealTimePredictor
 from sagemaker.model import Model
 from sagemaker.session import Session
@@ -23,11 +23,11 @@ class LDA(AmazonAlgorithmEstimatorBase):
 
     repo = 'lda:1'
 
-    num_topics = hp('num_topics', (isint, gt(0)), 'An integer greater than zero')
-    alpha0 = hp('alpha0', isnumber, "A float value")
-    max_restarts = hp('max_restarts', (isint, gt(0)), 'An integer greater than zero')
-    max_iterations = hp('max_iterations', (isint, gt(0)), 'An integer greater than zero')
-    tol = hp('tol', (isnumber, gt(0)), "A positive float")
+    num_topics = hp('num_topics', gt(0), 'An integer greater than zero', int)
+    alpha0 = hp('alpha0', (), "A float value", float)
+    max_restarts = hp('max_restarts', gt(0), 'An integer greater than zero', int)
+    max_iterations = hp('max_iterations', gt(0), 'An integer greater than zero', int)
+    tol = hp('tol', (gt(0)), "A positive float", float)
 
     def __init__(self, role, train_instance_type, num_topics,
                  alpha0=None, max_restarts=None, max_iterations=None, tol=None, **kwargs):